+++ src/Smartstore.Core/Platform/Messaging/Domain/EmailAccount.cs
@@ -1,13 +1,12 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Runtime.Serialization;
+using MailKit.Security;
 using Smartstore.Data.Caching;
 using Smartstore.Net.Mail;
 
 namespace Smartstore.Core.Messaging
 {
-    // TODO: (mg) (core) remove required attribute at EmailAccount.Username and EmailAccount.Password later (migration required).
-
     /// <summary>
     /// Represents an email account.
     /// </summary>
@@ -40,19 +39,32 @@ namespace Smartstore.Core.Messaging
         /// <summary>
         /// Gets or sets an email user name.
         /// </summary>
-        [Required, StringLength(255)]
+        [StringLength(255)]
         public string Username { get; set; }
 
         /// <summary>
         /// Gets or sets an email password.
         /// </summary>
-        [Required, StringLength(255)]
+        [StringLength(255)]
         public string Password { get; set; }
 
-        /// <summary>
-        /// Gets or sets a value that controls whether the SmtpClient uses Secure Sockets Layer (SSL) to encrypt the connection.
-        /// </summary>
+        [Obsolete("Use SecureOption instead.")]
         public bool EnableSsl { get; set; }
+
+        /// <summary>
+        /// Gets or sets the option identifier for SSL and/or TLS encryption to be used.
+        /// </summary>
+        public int SecureOptionId { get; set; }
+
+        /// <summary>
+        /// Gets or sets an option for SSL and/or TLS encryption to be used.
+        /// </summary>
+        [NotMapped]
+        public SecureSocketOptions SecureOption
+        {
+            get => (SecureSocketOptions)SecureOptionId;
+            set => SecureOptionId = (int)value;
+        }
 
         /// <summary>
         /// Gets or sets a value that controls whether the default system credentials of the application are sent with requests.