+++ changelog.md
@@ -7,6 +7,7 @@
 - **easyCredit** plugin (commercial)
 - **DependingPrices** plugin (commercial) Enables configuration of prices depending on customer groups, language, store or customer number.
 - New app system settings: `UsePooledDbContextFactory`, `UseDbCache`, `UseSequentialDbDataReader`
+- New option for e-mail accounts to configure SMTP connection encryption.
 
 ### Improvements